Tides v0.4 (#93)
* tides: initial release

This screen uses www.worldtides.info to get tide information about
a location.

This initial release is really simple and only display the current
time, previous tide and tide to come.

* tides: localize tides type

* tides: fix timezone handling

worldtides is not returning a correct timezone so we need to fix
that by removing the timezone and using a naive datetime instead.

* tides: handle switch

Refresh extreme when the date is reached.

* tides: use moving window for data

Refresh data when we don't get enough of them available. This way
we should always have the 7 incoming tides. This will allow to have
information about upcoming days.

* tides: fix Escalles coordinates

* tides: fix timezone computing

* tides: change page layout

* tides: add background image

* tides: improve second display on clock

* tides: add info on data and picture

* tides: fix color of labels

* tides: add location name

* tides: add display of tides to come

* tides: translate tides list

* tides: fix minute display in list

* tides: fix refresh strategy

Fix time comparison used to call a data refresh. This fixes
scrolling problem.

* tides: fix position of clock

* tides: fix display of next tides

Previous code was not correctly setting the time to the local
timezone.

* tides: better handling of non 200 error

* tides: fix error handling for answer

* tides: better error handling

This patch updates the tides screen logic to fail better if ever
there is a connection error at start. Switching to the screen will
trigger a data fetch if none has been succesfull till now.

* tides: gracefully fail if fetched data are invalid

* tides: fix error handling